requirements.txt
google-cloud-speech == 2.25.0
google-cloud-storage == 2.16.0
google-cloud-aiplatform == 1.44.0